requirements.txt
nose
django-toolbelt
djangorestframework
markdown
django-filter